+++ apps/benchmark/pages/button-variants/stitches.tsx
@@ -1,0 +1,19 @@
+import dynamic from "next/dynamic";
+
+const SierpinskiTriangle = () => {
+  const StitchesTest = dynamic(
+    () => import("../../bench/button-variants/stitches"),
+    {
+      ssr: false,
+    }
+  );
+
+  return (
+    <>
+      <h1 className="text-lg font-bold">stitches</h1>
+      <StitchesTest />
+    </>
+  );
+};
+
+export default SierpinskiTriangle;