+++ apps/benchmark/pages/button/baseline.tsx
@@ -1,0 +1,16 @@
+import dynamic from "next/dynamic";
+
+const SierpinskiTriangle = () => {
+  const StitchesTest = dynamic(() => import("../../bench/button/baseline"), {
+    ssr: false,
+  });
+
+  return (
+    <>
+      <h1 className="text-lg font-bold">Baseline</h1>
+      <StitchesTest />
+    </>
+  );
+};
+
+export default SierpinskiTriangle;