+++ apps/docs/components/ui/index.ts
@@ -2,3 +2,4 @@ export * from "./appbar";
 export * from "./button";
 export * from "./icon-button";
 export * from "./text";
+export * from "./layout";