+++ apps/docs/components/ui/text.tsx
@@ -13,12 +13,19 @@ const Text = classed("p", "leading-normal", {
       bold: "font-bold",
       medium: "font-medium",
     },
+    variant: {
+      title: "font-bold text-4xl",
+    },
 
     color: {
       primary: "text-radix-blue9",
       secondary: "text-radix-slate11",
       black: "text-black",
       white: "text-white",
+    },
+
+    center: {
+      true: "text-center",
     },
   },
 });