+++ apps/docs/pages/_app.tsx
@@ -1,5 +1,6 @@
 import "styles/globals.css";
 import "nextra-theme-docs/style.css";
+import "styles/theme-override.css";
 import type { AppProps } from "next/app";
 
 export default function Nextra({ Component, pageProps }: AppProps) {