+++ apps/docs/tailwind.config.js
@@ -1,9 +1,12 @@
+const plugin = require("tailwindcss/plugin");
+
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   darkMode: "class",
   content: [
     "./pages/**/*.{js,ts,jsx,tsx,md,mdx}",
     "./components/**/*.{js,ts,jsx,tsx,md,mdx}",
+    "./features/**/*.{js,ts,jsx,tsx,md,mdx}",
     "./theme.config.js",
   ],
   theme: {
@@ -25,5 +28,11 @@ module.exports = {
       colors: ["blue", "slate", "cyan", "violet", "green"],
     }),
     require("@tailwindcss/typography"),
+    plugin(function ({ addVariant }) {
+      addVariant(
+        "ios",
+        "@supports (selector(:nth-child(1 of x))) or (-webkit-touch-callout: none)"
+      );
+    }),
   ],
 };