+++ packages/react/index.ts
@@ -1,7 +1,9 @@
 export { classedProxy as classed } from "./src/proxy";
+export { createClassed } from "./src/classed";
 
 export type {
   InferVariantProps,
   ClassedComponentType,
   VariantProps,
+  DerivedComponentType,
 } from "./src/types";