+++ packages/react/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @tw-classed/react
 
+## 1.3.0-canary.4
+
+### Patch Changes
+
+- [`e7596bb9`](https://github.com/sannajammeh/tw-classed/commit/e7596bb9e0672ba22a83bcafae2e3925848c4596) Thanks [@sannajammeh](https://github.com/sannajammeh)! - Fixes "This module is a CJS module" in node.js with type:"module"
+
+- Updated dependencies [[`e7596bb9`](https://github.com/sannajammeh/tw-classed/commit/e7596bb9e0672ba22a83bcafae2e3925848c4596)]:
+  - @tw-classed/core@1.3.0-canary.4
+
+## 1.3.0-canary.3
+
+### Patch Changes
+
+- Updated dependencies [[`f328e321`](https://github.com/sannajammeh/tw-classed/commit/f328e32140a24d472f7b8d396181c36022afaebd)]:
+  - @tw-classed/core@1.3.0-canary.3
+
+## 1.3.0-canary.2
+
+### Patch Changes
+
+- [`29570c7e`](https://github.com/sannajammeh/tw-classed/commit/29570c7e185b2ee1b92259a3a97629ce7650e9dd) Thanks [@sannajammeh](https://github.com/sannajammeh)! - Export DerivedComponentType to handle as prop in derived components
+
+- Updated dependencies [[`29570c7e`](https://github.com/sannajammeh/tw-classed/commit/29570c7e185b2ee1b92259a3a97629ce7650e9dd)]:
+  - @tw-classed/core@1.3.0-canary.2
+
+## 1.3.0-canary.1
+
+### Patch Changes
+
+- [`79be37a2`](https://github.com/sannajammeh/tw-classed/commit/79be37a2543a46b85358f44f7ee8c27f729fb912) Thanks [@sannajammeh](https://github.com/sannajammeh)! - This release extends core to support the createClassed API, adds improvements to the documentation
+
+- Updated dependencies [[`79be37a2`](https://github.com/sannajammeh/tw-classed/commit/79be37a2543a46b85358f44f7ee8c27f729fb912)]:
+  - @tw-classed/core@1.3.0-canary.1
+
+## 1.3.0-canary.0
+
+### Minor Changes
+
+- [#52](https://github.com/sannajammeh/tw-classed/pull/52) [`47437bca`](https://github.com/sannajammeh/tw-classed/commit/47437bcab160071d863280a9f9541af6836a02b1) Thanks [@sannajammeh](https://github.com/sannajammeh)! - Add createClassed API to extend the classed function with additional functionality. Currently only "merger" prop is supported
+
 ## 1.2.5
 
 ### Patch Changes